Renamed table config
