+++ src/app/modules/cars/edit-table/table-config.ts
